refactor(workflow)!: reorganize workflow module structure
* Move workflow.ts from commands/ to dedicated workflow/ directory
* Update import paths in main.ts to reflect new structure
* fix(ui): rename sidebar button from "Cody Generate Reply" to "Cody Inference"

BREAKING CHANGE: workflow module path has changed from 'commands/workflow' to 'workflow/workflow'
