+++ vscode/src/main.ts
@@ -72,7 +72,7 @@ import { executeTestChatCommand } from './commands/execute/test-chat'
 import { CodySourceControl } from './commands/scm/source-control'
 import type { CodyCommandArgs } from './commands/types'
 import { newCodyCommandArgs } from './commands/utils/get-commands'
-import { registerWorkflowCommands } from './commands/workflow'
+import { registerWorkflowCommands } from './workflow/workflow'
 import { createInlineCompletionItemProvider } from './completions/create-inline-completion-item-provider'
 import { getConfiguration } from './configuration'
 import { exposeOpenCtxClient } from './context/openctx'
